+++ tests/api/test_account.py
@@ -1,11 +1,12 @@
 import json
 import re
-from unittest.mock import Mock, patch
+from unittest.mock import MagicMock, Mock, patch
 
 import graphene
 import pytest
 from django.contrib.auth import get_user_model
 from django.contrib.auth.tokens import default_token_generator
+from django.core.files import File
 from django.shortcuts import reverse
 
 from saleor.account.models import Address, User
@@ -15,8 +16,12 @@ from saleor.graphql.account.mutations import (
 from saleor.graphql.core.enums import PermissionEnum
 from saleor.order.models import FulfillmentStatus
 from tests.api.utils import get_graphql_content
-
-from .utils import assert_no_permission, convert_dict_keys_to_camel_case
+from tests.utils import create_image
+from .utils import (
+    assert_no_permission,
+    convert_dict_keys_to_camel_case,
+    get_multipart_request_body,
+)
 
 
 def test_create_token_mutation(admin_client, staff_user):
@@ -97,6 +102,11 @@ def test_query_user(
     user.default_shipping_address.save()
     user.addresses.add(address.get_copy())
 
+    avatar_mock = MagicMock(spec=File)
+    avatar_mock.name = 'image.jpg'
+    user.avatar = avatar_mock
+    user.save()
+
     query = """
     query User($id: ID!) {
         user(id: $id) {
@@ -149,6 +159,9 @@ def test_query_user(
                 isDefaultShippingAddress
                 isDefaultBillingAddress
             }
+            avatar {
+                url
+            }
         }
     }
     """
@@ -164,6 +177,7 @@ def test_query_user(
     assert data['isStaff'] == user.is_staff
     assert data['isActive'] == user.is_active
     assert data['orders']['totalCount'] == user.orders.count()
+    assert data['avatar']['url']
 
     assert len(data['addresses']) == user.addresses.count()
     for address in data['addresses']:
@@ -759,6 +773,9 @@ def test_staff_create(
                 permissions {
                     code
                 }
+                avatar {
+                    url
+                }
             }
         }
     }
@@ -778,6 +795,10 @@ def test_staff_create(
     assert data['user']['email'] == email
     assert data['user']['isStaff']
     assert data['user']['isActive']
+    assert re.match(
+        r'http://testserver/media/user-avatars/avatar\d+.*',
+        data['user']['avatar']['url']
+    )
     permissions = data['user']['permissions']
     assert permissions[0]['code'] == 'MANAGE_PRODUCTS'
 
@@ -1479,3 +1500,118 @@ def test_customer_change_default_address_invalid_address(
     assert (
         content['data']['customerSetDefaultAddress']['errors'][0]['field'] ==
         'id')
+
+
+USER_AVATAR_UPDATE_MUTATION = """
+    mutation userAvatarUpdate($image: Upload!) {
+        userAvatarUpdate(image: $image) {
+            user {
+                avatar {
+                    url
+                }
+            }
+        }
+    }
+"""
+
+
+def test_user_avatar_update_mutation_permission(api_client):
+    """ Should raise error if user is not staff. """
+
+    query = USER_AVATAR_UPDATE_MUTATION
+
+    image_file, image_name = create_image('avatar')
+    variables = {'image': image_name}
+    body = get_multipart_request_body(query, variables, image_file, image_name)
+    response = api_client.post_multipart(body)
+
+    assert_no_permission(response)
+
+def test_user_avatar_update_mutation(monkeypatch, staff_api_client):
+    query = USER_AVATAR_UPDATE_MUTATION
+
+    user = staff_api_client.user
+
+    mock_create_thumbnails = Mock(return_value=None)
+    monkeypatch.setattr(
+        ('saleor.graphql.account.mutations.'
+         'create_user_avatar_thumbnails.delay'),
+        mock_create_thumbnails)
+
+    image_file, image_name = create_image('avatar')
+    variables = {'image': image_name}
+    body = get_multipart_request_body(query, variables, image_file, image_name)
+    response = staff_api_client.post_multipart(body)
+    content = get_graphql_content(response)
+
+    data = content['data']['userAvatarUpdate']
+    user.refresh_from_db()
+
+    assert user.avatar
+    assert data['user']['avatar']['url'].startswith(
+        'http://testserver/media/user-avatars/avatar'
+    )
+
+    # The image creation should have triggered a warm-up
+    mock_create_thumbnails.assert_called_once_with(user_id=user.pk)
+
+
+def test_user_avatar_update_mutation_image_exists(staff_api_client):
+    query = USER_AVATAR_UPDATE_MUTATION
+
+    user = staff_api_client.user
+    avatar_mock = MagicMock(spec=File)
+    avatar_mock.name = 'image.jpg'
+    user.avatar = avatar_mock
+    user.save()
+
+    image_file, image_name = create_image('new_image')
+    variables = {'image': image_name}
+    body = get_multipart_request_body(query, variables, image_file, image_name)
+    response = staff_api_client.post_multipart(body)
+    content = get_graphql_content(response)
+
+    data = content['data']['userAvatarUpdate']
+    user.refresh_from_db()
+
+    assert user.avatar != avatar_mock
+    assert data['user']['avatar']['url'].startswith(
+        'http://testserver/media/user-avatars/new_image'
+    )
+
+
+USER_AVATAR_DELETE_MUTATION = """
+    mutation userAvatarDelete {
+        userAvatarDelete {
+            user {
+                avatar {
+                    url
+                }
+            }
+        }
+    }
+"""
+
+
+def test_user_avatar_delete_mutation_permission(api_client):
+    """ Should raise error if user is not staff. """
+
+    query = USER_AVATAR_DELETE_MUTATION
+
+    response = api_client.post_graphql(query)
+
+    assert_no_permission(response)
+
+
+def test_user_avatar_delete_mutation(staff_api_client):
+    query = USER_AVATAR_DELETE_MUTATION
+
+    user = staff_api_client.user
+
+    response = staff_api_client.post_graphql(query)
+    content = get_graphql_content(response)
+
+    user.refresh_from_db()
+
+    assert not user.avatar
+    assert not content['data']['userAvatarDelete']['user']['avatar']